System User Handling (#121)
* System User Handling

* Validation Fixes

* Commented Test Case handled

* Constraint changes

+++ src/main/kotlin/com/hypto/iam/server/service/UsersService.kt
@@ -7,20 +7,20 @@ import com.hypto.iam.server.configs.AppConfig
 import com.hypto.iam.server.db.repositories.OrganizationRepo
 import com.hypto.iam.server.db.repositories.PasscodeRepo
 import com.hypto.iam.server.db.repositories.UserRepo
-import com.hypto.iam.server.db.tables.pojos.Users
+import com.hypto.iam.server.db.tables.records.UsersRecord
 import com.hypto.iam.server.exceptions.EntityNotFoundException
 import com.hypto.iam.server.exceptions.InternalException
+import com.hypto.iam.server.extensions.PaginationContext
 import com.hypto.iam.server.extensions.toUserStatus
 import com.hypto.iam.server.idp.IdentityGroup
 import com.hypto.iam.server.idp.IdentityProvider
-import com.hypto.iam.server.idp.NextToken
 import com.hypto.iam.server.idp.PasswordCredentials
 import com.hypto.iam.server.idp.RequestContext
 import com.hypto.iam.server.idp.UserAlreadyExistException
 import com.hypto.iam.server.models.BaseSuccessResponse
-import com.hypto.iam.server.models.PaginationOptions
 import com.hypto.iam.server.models.UpdateUserRequest
 import com.hypto.iam.server.models.User
+import com.hypto.iam.server.models.UserPaginatedResponse
 import com.hypto.iam.server.models.VerifyEmailRequest
 import com.hypto.iam.server.security.AuthenticationException
 import com.hypto.iam.server.utils.IamResources
@@ -41,7 +41,13 @@ class UsersServiceImpl : KoinComponent, UsersService {
 
     override suspend fun createUser(
         organizationId: String,
-        credentials: PasswordCredentials,
+        loginAccess: Boolean,
+        username: String,
+        preferredUsername: String?,
+        name: String?,
+        email: String?,
+        phoneNumber: String?,
+        password: String?,
         createdBy: String?,
         verified: Boolean
     ): User {
@@ -49,60 +55,82 @@ class UsersServiceImpl : KoinComponent, UsersService {
             ?: throw EntityNotFoundException("Invalid organization id name. Unable to create a user")
 
         if (userRepo.existsByAliasUsername(
-                credentials.preferredUsername,
-                credentials.email,
+                preferredUsername,
+                email,
                 organizationId,
                 appConfig.app.uniqueUsersAcrossOrganizations
             )
         )
             throw UserAlreadyExistException(
-                "Email - ${credentials.email} " +
-                    credentials.preferredUsername?.let {
-                        "or Username - ${credentials.preferredUsername}"
-                    }.orEmpty() +
+                email?.let { "Email - $email" }.orEmpty() +
+                    preferredUsername?.let { "Username - $preferredUsername" }.orEmpty() +
                     "already registered. Unable to create user"
             )
 
-        val identityGroup = gson.fromJson(org.metadata.data(), IdentityGroup::class.java)
-        val user = identityProvider.createUser(
-            RequestContext(organizationId = organizationId, requestedPrincipal = createdBy ?: "unknown user", verified),
-            identityGroup, credentials
-        )
-        val userHrn = ResourceHrn(organizationId, "", IamResources.USER, user.username)
-        userRepo.insert(
-            Users(
-                userHrn.toString(), credentials.email, User.Status.enabled.value,
-                organizationId, LocalDateTime.now(), LocalDateTime.now(), verified, false, credentials.preferredUsername
+        if (loginAccess) {
+            val passwordCredentials = PasswordCredentials(
+                username = username,
+                preferredUsername = preferredUsername,
+                name = name,
+                email = email!!,
+                phoneNumber = phoneNumber ?: "",
+                password = password!!
             )
-        )
-        return getUser(userHrn, user)
+            val identityGroup = gson.fromJson(org.metadata.data(), IdentityGroup::class.java)
+            identityProvider.createUser(
+                RequestContext(
+                    organizationId = organizationId,
+                    requestedPrincipal = createdBy ?: "unknown user", verified
+                ),
+                identityGroup, passwordCredentials
+            )
+        }
+        val userHrn = ResourceHrn(organizationId, "", IamResources.USER, username)
+
+        val userRecord = userRepo.insert(
+            UsersRecord().apply {
+                this.hrn = userHrn.toString()
+                this.organizationId = organizationId
+                this.email = email
+                this.status = User.Status.enabled.value
+                this.verified = verified
+                this.deleted = false
+                this.createdAt = LocalDateTime.now()
+                this.updatedAt = LocalDateTime.now()
+                this.preferredUsername = preferredUsername
+                this.loginAccess = loginAccess
+                this.name = name
+                this.createdBy = createdBy
+            }
+        ) ?: throw InternalException("Unable to create user")
+        return getUser(userHrn, userRecord)
     }
 
     override suspend fun getUser(organizationId: String, userName: String): User {
-        val org = organizationRepo.findById(organizationId)
-            ?: throw EntityNotFoundException("Invalid organization id name. Unable to get user")
+        organizationRepo.findById(organizationId)
+            ?: throw EntityNotFoundException("Invalid organization id. Unable to get user")
         val userHrn = ResourceHrn(organizationId, "", IamResources.USER, userName)
-        val identityGroup = gson.fromJson(org.metadata.data(), IdentityGroup::class.java)
-        val user = identityProvider.getUser(identityGroup, userName)
-        return getUser(userHrn, user)
+        val userRecord = userRepo.findByHrn(userHrn.toString())
+            ?: throw EntityNotFoundException("Unable to find user")
+        return getUser(userHrn, userRecord)
     }
 
     override suspend fun getUserByEmail(
         organizationId: String?,
         email: String
     ): User {
-        val orgId = if (appConfig.app.uniqueUsersAcrossOrganizations) {
-            val userRecord = userRepo.findByEmail(email)
+        val userRecord = if (appConfig.app.uniqueUsersAcrossOrganizations) {
+            val user = userRepo.findByEmail(email)
                 ?: throw EntityNotFoundException("User with email - $email not found")
-            userRecord.organizationId
+            organizationRepo.findById(user.organizationId)
+                ?: throw EntityNotFoundException("Invalid organization id name. Unable to get user")
+            user
         } else {
             organizationId ?: throw EntityNotFoundException("Organization id is required")
-        }
-        val org = organizationRepo.findById(orgId)
-            ?: throw EntityNotFoundException("Invalid organization id name. Unable to get user")
-        val identityGroup = gson.fromJson(org.metadata.data(), IdentityGroup::class.java)
-        val user = identityProvider.getUserByEmail(identityGroup, email)
-        return getUser(ResourceHrn(orgId, "", IamResources.USER, user.username), user)
+            userRepo.findByEmail(email, organizationId)
+        } ?: throw EntityNotFoundException("User with email - $email not found")
+
+        return getUser(ResourceHrn(userRecord.hrn), userRecord)
     }
 
     override suspend fun changeUserPassword(
@@ -113,6 +141,13 @@ class UsersServiceImpl : KoinComponent, UsersService {
     ): BaseSuccessResponse {
         val org = organizationRepo.findById(organizationId)
             ?: throw EntityNotFoundException("Invalid organization id name. Unable to authenticate user")
+
+        val userHrn = ResourceHrn(organizationId, "", IamResources.USER, userName)
+        val userRecord = userRepo.findByHrn(userHrn.toString())
+            ?: throw EntityNotFoundException("Unable to find user")
+        if (!userRecord.loginAccess)
+            throw BadRequestException("User - $userName does not have login access")
+
         val identityGroup = gson.fromJson(org.metadata.data(), IdentityGroup::class.java)
         identityProvider.authenticate(identityGroup, userName, oldPassword)
         identityProvider.setUserPassword(identityGroup, userName, newPassword)
@@ -127,45 +162,50 @@ class UsersServiceImpl : KoinComponent, UsersService {
     ): BaseSuccessResponse {
         val org = organizationRepo.findById(organizationId)
             ?: throw EntityNotFoundException("Invalid organization id name. Unable to set user password")
+
+        val userHrn = ResourceHrn(user.hrn)
+        val userRecord = userRepo.findByHrn(userHrn.toString())
+            ?: throw EntityNotFoundException("Unable to find user")
+        if (!userRecord.loginAccess)
+            throw BadRequestException("User - ${userHrn.resourceInstance} does not have login access")
+
         val identityGroup = gson.fromJson(org.metadata.data(), IdentityGroup::class.java)
         identityProvider.setUserPassword(identityGroup, user.username, password)
-        passcodeRepo.deleteByEmailAndPurpose(user.email, VerifyEmailRequest.Purpose.reset)
+        passcodeRepo.deleteByEmailAndPurpose(user.email!!, VerifyEmailRequest.Purpose.reset)
         return BaseSuccessResponse(true)
     }
 
     override suspend fun listUsers(
         organizationId: String,
-        nextToken: String?,
-        pageSize: Int?
-    ): Triple<List<User>, NextToken?, PaginationOptions> {
+        paginationContext: PaginationContext
+    ): UserPaginatedResponse {
         val org = organizationRepo.findById(organizationId)
             ?: throw EntityNotFoundException("Invalid organization id name. Unable to get user")
-        val identityGroup = gson.fromJson(org.metadata.data(), IdentityGroup::class.java)
-        val (users, nextToken) = identityProvider.listUsers(
-            identityGroup = identityGroup,
-            pageToken = nextToken, limit = pageSize
-        )
-        val pageContext = PaginationOptions(pageSize)
-        val externalUserTypeUsers = users.map { user ->
-            val userHrn = ResourceHrn(organizationId, "", IamResources.USER, user.username)
-            getUser(userHrn, user)
+
+        val users = userRepo.fetchUsers(organizationId, paginationContext).map { user ->
+            getUser(ResourceHrn(user.hrn), user)
         }.toList()
-        return Triple(externalUserTypeUsers, nextToken, pageContext)
+
+        val newContext = PaginationContext.from(users.lastOrNull()?.hrn, paginationContext)
+        return UserPaginatedResponse(
+            users,
+            newContext.nextToken,
+            newContext.toOptions()
+        )
     }
 
     private fun getUser(
         userHrn: ResourceHrn,
-        user: com.hypto.iam.server.idp.User
+        user: UsersRecord
     ) = User(
         hrn = userHrn.toString(),
-        username = user.username,
+        username = userHrn.resourceInstance.toString(),
         preferredUsername = user.preferredUsername,
         name = user.name,
         organizationId = userHrn.organization,
         email = user.email,
-        status = if (user.isEnabled) User.Status.enabled else User.Status.disabled,
+        status = if (user.status == User.Status.enabled.value) User.Status.enabled else User.Status.disabled,
         verified = user.verified,
-        phone = user.phoneNumber,
         createdBy = user.createdBy,
         loginAccess = user.loginAccess
     )
@@ -181,21 +221,26 @@ class UsersServiceImpl : KoinComponent, UsersService {
         val org = organizationRepo.findById(organizationId)
             ?: throw EntityNotFoundException("Invalid organization id name. Unable to get user")
         val userHrn = ResourceHrn(organizationId, "", IamResources.USER, userName)
-        val identityGroup = gson.fromJson(org.metadata.data(), IdentityGroup::class.java)
-        val userStatus = status?.toUserStatus()
-        val user = identityProvider.updateUser(
-            identityGroup = identityGroup,
-            name = name,
-            userName = userName,
-            phone = phone,
-            status = userStatus,
-            verified = verified
-        )
-
-        userRepo.update(userHrn.toString(), userStatus, verified)
+        val userRecord = userRepo.findByHrn(userHrn.toString())
             ?: throw EntityNotFoundException("User not found")
 
-        return getUser(userHrn, user)
+        val userStatus = status?.toUserStatus()
+        if (userRecord.loginAccess) {
+            val identityGroup = gson.fromJson(org.metadata.data(), IdentityGroup::class.java)
+
+            identityProvider.updateUser(
+                identityGroup = identityGroup,
+                name = name,
+                userName = userName,
+                phone = phone,
+                status = userStatus,
+                verified = verified
+            )
+        }
+        val updatedUserRecord = userRepo.update(userHrn.toString(), userStatus, verified, name)
+            ?: throw EntityNotFoundException("User not found")
+
+        return getUser(userHrn, updatedUserRecord)
     }
 
     override suspend fun deleteUser(organizationId: String, userName: String): BaseSuccessResponse {
@@ -205,10 +250,14 @@ class UsersServiceImpl : KoinComponent, UsersService {
         if (org.rootUserHrn == userHrn.toString())
             throw BadRequestException("Cannot delete Root User")
 
-        val identityGroup = gson.fromJson(org.metadata.data(), IdentityGroup::class.java)
-        identityProvider.deleteUser(identityGroup, userName)
+        val userRecord = userRepo.findByHrn(userHrn.toString())
+            ?: throw EntityNotFoundException("User not found")
 
-        userRepo.delete(userHrn.toString()) ?: throw EntityNotFoundException("User not found")
+        if (userRecord.loginAccess) {
+            val identityGroup = gson.fromJson(org.metadata.data(), IdentityGroup::class.java)
+            identityProvider.deleteUser(identityGroup, userName)
+        }
+        userRepo.delete(userHrn.toString())
 
         return BaseSuccessResponse(true)
     }
@@ -216,10 +265,16 @@ class UsersServiceImpl : KoinComponent, UsersService {
     override suspend fun authenticate(organizationId: String, userName: String, password: String): User {
         val org = organizationRepo.findById(organizationId)
             ?: throw EntityNotFoundException("Invalid organization id name. Unable to authenticate user")
+
+        val userRecord = userRepo.findByEmail(userName, organizationId)
+            ?: throw EntityNotFoundException("User not found")
+        if (!userRecord.loginAccess)
+            throw BadRequestException("User - $userName does not have login access")
+
         val identityGroup = gson.fromJson(org.metadata.data(), IdentityGroup::class.java)
-        val user = identityProvider.authenticate(identityGroup, userName, password)
-        val userHrn = ResourceHrn(organizationId, "", IamResources.USER, userName)
-        return getUser(userHrn, user)
+        identityProvider.authenticate(identityGroup, userName, password)
+
+        return getUser(ResourceHrn(userRecord.hrn), userRecord)
     }
 
     override suspend fun authenticate(username: String, password: String): User {
@@ -232,10 +287,13 @@ class UsersServiceImpl : KoinComponent, UsersService {
         val org = organizationRepo.findById(userRecord.organizationId)
             ?: throw InternalException("Internal error while trying to authenticate the identity.")
 
+        if (!userRecord.loginAccess)
+            throw BadRequestException("User - $username does not have login access")
+
         val identityGroup = gson.fromJson(org.metadata.data(), IdentityGroup::class.java)
         val user = identityProvider.authenticate(identityGroup, username, password)
         val userHrn = ResourceHrn(userRecord.organizationId, "", IamResources.USER, user.username)
-        return getUser(userHrn, user)
+        return getUser(userHrn, userRecord)
     }
 }
 
@@ -245,14 +303,20 @@ class UsersServiceImpl : KoinComponent, UsersService {
 interface UsersService {
     suspend fun createUser(
         organizationId: String,
-        credentials: PasswordCredentials,
+        loginAccess: Boolean,
+        username: String,
+        preferredUsername: String?,
+        name: String?,
+        email: String?,
+        phoneNumber: String?,
+        password: String?,
         createdBy: String?,
         verified: Boolean
     ): User
 
     suspend fun getUser(organizationId: String, userName: String): User
-    suspend fun listUsers(organizationId: String, nextToken: String?, pageSize: Int?):
-        Triple<List<User>, NextToken?, PaginationOptions>
+    suspend fun listUsers(organizationId: String, paginationContext: PaginationContext):
+        UserPaginatedResponse
 
     suspend fun updateUser(
         organizationId: String,
